got rid of compiler warnings

+++ linux/GlxContext.cpp
@@ -259,8 +259,8 @@ GlContext::createContext(ContextSettings& settings, GlContext* share) {
 				// create the context
 				int attributes[] = {
 
-					GLX_CONTEXT_MAJOR_VERSION_ARB, _settings.majorVersion,
-					GLX_CONTEXT_MINOR_VERSION_ARB, _settings.minorVersion,
+					GLX_CONTEXT_MAJOR_VERSION_ARB, (int)_settings.majorVersion,
+					GLX_CONTEXT_MINOR_VERSION_ARB, (int)_settings.minorVersion,
 					GLX_CONTEXT_PROFILE_MASK_ARB, GLX_CONTEXT_COMPATIBILITY_PROFILE_BIT_ARB,
 					0, 0
 				};
